UI System needs a strong refactory. There should only be widget, no container.

+++ src/main/java/theking530/staticcore/utilities/RenderingUtilities.java
@@ -4,7 +4,7 @@ import com.mojang.blaze3d.systems.RenderSystem;
 
 public class RenderingUtilities {
 	public static void applyScissorMask(RectangleBounds bounds) {
-		RenderSystem.enableScissor(bounds.getX(), bounds.getY(), bounds.getWidth(), bounds.getHeight());
+		RenderSystem.enableScissor((int) bounds.getX(), (int) bounds.getY(), bounds.getWidth(), bounds.getHeight());
 	}
 
 	public static void clearScissorMask() {
